requirements.txt
alembic==1.0.7
bcrypt==3.1.6
blinker==1.4
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
Pillow==5.4.1
psycopg2==2.8.1
pycparser==2.19
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
sqlparse==0.3.0
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
WTForms==2.2.1